fix eas build
