+++ starters/next-expo-solito/apps/expo/package.json
@@ -21,7 +21,6 @@
     "expo-constants": "^14.2.1",
     "expo-dev-client": "^2.1.5",
     "expo-font": "^11.1.1",
-    "expo-image": "^1.0.0",
     "expo-linear-gradient": "^12.1.2",
     "expo-linking": "^4.0.1",
     "expo-router": "^1.4.3",
